requirements-clean.txt
# Research Assistant Platform - Clean Dependencies
# Only includes libraries that are actually used in the application

# ==============================================================================
# CORE FRAMEWORK
# ==============================================================================
streamlit>=1.30.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0

# ==============================================================================
# AUTHENTICATION & NAVIGATION
# ==============================================================================
st-pages>=0.4.0,<1.0.0
streamlit-authenticator>=0.4.2,<1.0.0
pyyaml>=6.0,<7.0.0

# ==============================================================================
# AI & LLM - CORE
# ==============================================================================
openai>=1.0.0,<2.0.0
langchain>=0.1.0,<0.3.0
langchain-openai>=0.0.5,<0.3.0
langchain-community>=0.0.20,<0.3.0
langchain-chroma>=0.1.0,<0.2.0
langchain-text-splitters>=0.0.1,<0.3.0
langchainhub>=0.1.14,<0.2.0
langchain-core>=0.1.0,<0.3.0

# ==============================================================================
# SEARCH INTEGRATIONS
# ==============================================================================
# Google Search (Optional - for Google Scholar)
langchain-google-community>=0.0.1,<0.2.0

# ArXiv
arxiv>=2.1.0,<3.0.0

# DuckDuckGo
duckduckgo-search>=4.0.0,<7.0.0

# ==============================================================================
# DOCUMENT PROCESSING
# ==============================================================================
pymupdf>=1.23.0,<2.0.0              # PDF processing (fitz)
beautifulsoup4>=4.12.0,<5.0.0       # HTML parsing
requests>=2.31.0,<3.0.0             # HTTP requests

# ==============================================================================
# VECTOR STORE & EMBEDDINGS
# ==============================================================================
chromadb>=0.4.0,<0.6.0

# ==============================================================================
# DATABASE (Optional - for prompt management)
# ==============================================================================
pymongo>=4.6.0,<5.0.0               # MongoDB client

# ==============================================================================
# UTILITIES
# ==============================================================================
tiktoken>=0.5.0,<0.8.0              # Token counting

# ==============================================================================
# DEVELOPMENT & TESTING (Optional)
# ==============================================================================
# Uncomment for development
# pytest>=7.4.0,<9.0.0
# pytest-cov>=4.1.0,<6.0.0
# black>=23.0.0,<25.0.0
# flake8>=6.0.0,<8.0.0
